// interfaces/user.ts
// Generated by https://quicktype.io

export interface User {
  id_user:       number;
  id_role:         number;
  name:         string;
  email:        string;
  password?:     string;
  phone:        string;
  address:      string;
  postal_code:  string;
  image_url:      string;
  created_at?:   string;
  updated_at?:   string;
  status?:       boolean;
}